+++ src/test/ExSln2/c99/gen/hal_LedArray_LedArrayStatus.h
@@ -1,0 +1,6 @@
+// finlang generated file for c# hal.LedArray.LedArrayStatus type
+#pragma once
+
+
+
+    typedef enum hal_LedArray_LedArrayStatus { hal_LedArray_LedArrayStatus_OFF, hal_LedArray_LedArrayStatus_ON, hal_LedArray_LedArrayStatus_BLINKING } hal_LedArray_LedArrayStatus;