+++ src/test/ExSln2/c99/gen/hal_LedArray_LedInfo.c
@@ -1,0 +1,5 @@
+// finlang generated file for c# hal.LedArray.LedInfo type
+
+#include "hal_LedArray_LedInfo.h"
+
+