+++ src/test/ExSln2/c99/gen/hal_LedArray_LedInfo.h
@@ -1,0 +1,12 @@
+// finlang generated file for c# hal.LedArray.LedInfo type
+#pragma once
+
+#include <stdint.h>
+
+
+    typedef struct hal_LedArray_LedInfo hal_LedArray_LedInfo;
+struct hal_LedArray_LedInfo
+{
+        uint8_t something;
+};
+